Pull URL handling out of AppDelegate
