+++ HomeAssistant/AppDelegate.swift
@@ -47,22 +47,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     let sceneManager = SceneManager()
 
-    var webViewWindowControllerPromise: Promise<WebViewWindowController> {
-        if #available(iOS 13, *) {
-            return firstly { () -> Promise<WebViewSceneDelegate> in
-                sceneManager.scene(for: .init(activity: .webView))
-            }.compactMap { delegate in
-                delegate.windowController
-            }
-        } else {
-            return Promise(iOS12WindowControllerPromise)
-        }
-    }
-
-    var iOS12WindowController: WebViewWindowController?
-    private let iOS12WindowControllerPromise: Guarantee<WebViewWindowController>
-    private let iOS12WindowControllerSeal: (WebViewWindowController) -> Void
-
     private var zoneManager: ZoneManager?
 
     private var periodicUpdateTimer: Timer? {
@@ -74,7 +58,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     override init() {
-        (self.iOS12WindowControllerPromise, self.iOS12WindowControllerSeal) = Guarantee<WebViewWindowController>.pending()
+        
 
         super.init()
     }
@@ -115,8 +99,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                                 type: .unknown)
         Current.clientEventStore.addEvent(event)
 
-        self.registerCallbackURLKitHandlers()
-
         self.zoneManager = ZoneManager()
 
         UIApplication.shared.setMinimumBackgroundFetchInterval(UIApplication.backgroundFetchIntervalMinimum)
@@ -131,7 +113,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         } else {
             // window must be created before willFinishLaunching completes, or state restoration will not occur
-            setupWindowForiOS12()
+            sceneManager.compatibility.start()
         }
 
         return true
@@ -148,7 +130,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         if #available(iOS 13, *) {
 
         } else {
-            iOS12WindowControllerPromise.done { $0.setup() }
+            sceneManager.compatibility.setup()
         }
 
         _ = HomeAssistantAPI.authenticatedAPI()?.CreateEvent(eventType: "ios.finished_launching", eventData: [:])
@@ -158,7 +140,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         checkForUpdate()
 
         DispatchQueue.main.asyncAfter(deadline: .now() + .seconds(5)) {
-            self.webViewWindowControllerPromise.done {
+            self.sceneManager.webViewWindowControllerPromise.done {
                 $0.open(urlString: "/lovelace-teest/0")
             }
         }
@@ -191,17 +173,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             controller.addAction(UIAlertAction(title: "Ask Again Later", style: .cancel, handler: nil))
             window?.rootViewController?.present(controller, animated: true, completion: nil)
         }
-    }
-
-    func setupWindowForiOS12() {
-        if #available(iOS 13, *) {
-            fatalError("bad code path: iOS 12 window controller shouldn't be set up on iOS 13 or above")
-        }
-
-        let window = WebViewWindowController.window(foriOS12: ())
-        let windowController = WebViewWindowController(window: window)
-        self.iOS12WindowController = windowController
-        iOS12WindowControllerSeal(windowController)
     }
 
     func setupTokens() {
@@ -295,7 +266,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func application(_ application: UIApplication, shouldRestoreSecureApplicationState coder: NSCoder) -> Bool {
-        if iOS12WindowController?.requiresOnboarding == true {
+        if sceneManager.compatibility.windowController?.requiresOnboarding == true {
             Current.Log.info("disallowing state to be restored due to onboarding")
             return false
         }
@@ -329,7 +300,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         viewControllerWithRestorationIdentifierPath identifierComponents: [String],
         coder: NSCoder
     ) -> UIViewController? {
-        return iOS12WindowController?.viewController(withRestorationIdentifierPath: identifierComponents)
+        return sceneManager.compatibility.windowController?.viewController(withRestorationIdentifierPath: identifierComponents)
     }
 
     func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
@@ -427,33 +398,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
     }
 
-    func application(_ app: UIApplication,
-                     open url: URL, options: [UIApplication.OpenURLOptionsKey: Any] = [:]) -> Bool {
-        Current.Log.verbose("Received URL: \(url)")
-        var serviceData: [String: String] = [:]
-        if let queryItems = url.queryItems {
-            serviceData = queryItems
-        }
-        guard let host = url.host else { return true }
-        switch host.lowercased() {
-        case "x-callback-url":
-            return Manager.shared.handleOpen(url: url)
-        case "call_service":
-            callServiceURLHandler(url, serviceData)
-        case "fire_event":
-            fireEventURLHandler(url, serviceData)
-        case "send_location":
-            sendLocationURLHandler()
-        case "perform_action":
-            performActionURLHandler(url, serviceData: serviceData)
-        case "auth-callback": // homeassistant://auth-callback
-           NotificationCenter.default.post(name: Notification.Name("AuthCallback"), object: nil,
-                                           userInfo: ["url": url])
-        default:
-            Current.Log.warning("Can't route incoming URL: \(url)")
-            showAlert(title: L10n.errorLabel, message: L10n.UrlHandler.NoService.message(url.host!))
-        }
-        return true
+    func application(
+        _ app: UIApplication,
+        open url: URL,
+        options: [UIApplication.OpenURLOptionsKey: Any] = [:]
+    ) -> Bool {
+        return sceneManager.compatibility.urlHandler?.handle(url: url) ?? false
     }
 
     func application(_ application: UIApplication, performActionFor shortcutItem: UIApplicationShortcutItem,
@@ -477,28 +427,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
     }
 
-    func application(_ application: UIApplication, continue userActivity: NSUserActivity,
-                     restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
+    func application(
+        _ application: UIApplication,
+        continue userActivity: NSUserActivity,
+        restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void
+    ) -> Bool {
         Current.Log.info(userActivity)
-
-        switch Current.tags.handle(userActivity: userActivity) {
-        case .handled(let type):
-            let (icon, text) = { () -> (MaterialDesignIcons, String) in
-                 switch type {
-                 case .nfc:
-                     return (.nfcVariantIcon, L10n.Nfc.tagRead)
-                 case .generic:
-                     return (.qrcodeIcon, L10n.Nfc.genericTagRead)
-                 }
-             }()
-            showFullScreenConfirm(icon: icon, text: text)
-            return true
-        case .unhandled:
-            return false
-        case .open(let url):
-            // NFC-based URL
-            return self.application(application, open: url)
-        }
+        return sceneManager.compatibility.urlHandler?.handle(userActivity: userActivity) ?? false
     }
 
     func application(
@@ -557,175 +492,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }.finally {
             self.schedulePeriodicUpdateTimer()
         }
-    }
-
-    // swiftlint:disable:next function_body_length
-    private func registerCallbackURLKitHandlers() {
-        Manager.shared.callbackURLScheme = Manager.urlSchemes?.first
-
-        Manager.shared["fire_event"] = { parameters, success, failure, cancel in
-            guard let eventName = parameters["eventName"] else {
-                failure(XCallbackError.eventNameMissing)
-                return
-            }
-
-            var cleanParamters = parameters
-            cleanParamters.removeValue(forKey: "eventName")
-            let eventData = cleanParamters
-
-            _ = firstly {
-                HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.CreateEvent(eventType: eventName, eventData: eventData)
-            }.done { _ in
-                success(nil)
-            }.catch { error -> Void in
-                Current.Log.error("Received error from createEvent during X-Callback-URL call: \(error)")
-                failure(XCallbackError.generalError)
-            }
-        }
-
-        Manager.shared["call_service"] = { parameters, success, failure, cancel in
-            guard let service = parameters["service"] else {
-                failure(XCallbackError.serviceMissing)
-                return
-            }
-
-            let splitService = service.components(separatedBy: ".")
-            let serviceDomain = splitService[0]
-            let serviceName = splitService[1]
-
-            var cleanParamters = parameters
-            cleanParamters.removeValue(forKey: "service")
-            let serviceData = cleanParamters
-
-            _ = firstly {
-                HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.CallService(domain: serviceDomain, service: serviceName, serviceData: serviceData)
-            }.done { _ in
-                success(nil)
-            }.catch { error in
-                Current.Log.error("Received error from callService during X-Callback-URL call: \(error)")
-                failure(XCallbackError.generalError)
-            }
-        }
-
-        Manager.shared["send_location"] = { parameters, success, failure, cancel in
-            _ = firstly {
-                HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.GetAndSendLocation(trigger: .XCallbackURL)
-            }.done { _ in
-                success(nil)
-            }.catch { error in
-                Current.Log.error("Received error from getAndSendLocation during X-Callback-URL call: \(error)")
-                failure(XCallbackError.generalError)
-            }
-        }
-
-        Manager.shared["render_template"] = { parameters, success, failure, cancel in
-            guard let template = parameters["template"] else {
-                failure(XCallbackError.templateMissing)
-                return
-            }
-
-            var cleanParamters = parameters
-            cleanParamters.removeValue(forKey: "template")
-            let variablesDict = cleanParamters
-
-            _ = firstly {
-                HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.RenderTemplate(templateStr: template, variables: variablesDict)
-            }.done { rendered in
-                success(["rendered": rendered])
-            }.catch { error in
-                Current.Log.error("Received error from RenderTemplate during X-Callback-URL call: \(error)")
-                failure(XCallbackError.generalError)
-            }
-        }
-    }
-
-    private func fireEventURLHandler(_ url: URL, _ serviceData: [String: String]) {
-        // homeassistant://fire_event/custom_event?entity_id=device_tracker.entity
-
-        _ = firstly {
-            HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.CreateEvent(eventType: url.pathComponents[1], eventData: serviceData)
-            }.done { _ in
-                showAlert(title: L10n.UrlHandler.FireEvent.Success.title,
-                          message: L10n.UrlHandler.FireEvent.Success.message(url.pathComponents[1]))
-            }.catch { error -> Void in
-                showAlert(title: L10n.errorLabel,
-                          message: L10n.UrlHandler.FireEvent.Error.message(url.pathComponents[1],
-                                                                           error.localizedDescription))
-        }
-    }
-
-    private func callServiceURLHandler(_ url: URL, _ serviceData: [String: String]) {
-        // homeassistant://call_service/device_tracker.see?entity_id=device_tracker.entity
-        let domain = url.pathComponents[1].components(separatedBy: ".")[0]
-        let service = url.pathComponents[1].components(separatedBy: ".")[1]
-
-        _ = firstly {
-            HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.CallService(domain: domain, service: service, serviceData: serviceData)
-            }.done { _ in
-                showAlert(title: L10n.UrlHandler.CallService.Success.title,
-                          message: L10n.UrlHandler.CallService.Success.message(url.pathComponents[1]))
-            }.catch { error in
-                showAlert(title: L10n.errorLabel,
-                          message: L10n.UrlHandler.CallService.Error.message(url.pathComponents[1],
-                                                                             error.localizedDescription))
-        }
-    }
-
-    private func sendLocationURLHandler() {
-        // homeassistant://send_location/
-        _ = firstly {
-            HomeAssistantAPI.authenticatedAPIPromise
-            }.then { api in
-                api.GetAndSendLocation(trigger: .URLScheme)
-            }.done { _ in
-                showAlert(title: L10n.UrlHandler.SendLocation.Success.title,
-                          message: L10n.UrlHandler.SendLocation.Success.message)
-            }.catch { error in
-                showAlert(title: L10n.errorLabel,
-                          message: L10n.UrlHandler.SendLocation.Error.message(error.localizedDescription))
-        }
-    }
-
-    private func performActionURLHandler(_ url: URL, serviceData: [String: String]) {
-        let pathComponents = url.pathComponents
-        guard pathComponents.count > 1 else {
-            Current.Log.error("not enough path components for perform action handler")
-            return
-        }
-
-        let source: HomeAssistantAPI.ActionSource = {
-            if let sourceString = serviceData["source"],
-               let source = HomeAssistantAPI.ActionSource(rawValue: sourceString) {
-                return source
-            } else {
-                return .URLHandler
-            }
-        }()
-
-        let actionID = url.pathComponents[1]
-
-        guard let action = Current.realm().object(ofType: Action.self, forPrimaryKey: actionID) else {
-            showFullScreenConfirm(icon: .alertCircleIcon, text: L10n.UrlHandler.Error.actionNotFound)
-            return
-        }
-
-        showFullScreenConfirm(icon: MaterialDesignIcons(named: action.IconName), text: action.Text)
-
-        HomeAssistantAPI.authenticatedAPI()?
-            .HandleAction(actionID: actionID, source: source)
-            .cauterize()
     }
 
     func checkForUpdate() {
@@ -1090,7 +856,7 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
         }
 
         if let openURLRaw = userInfo["url"] as? String {
-            webViewWindowControllerPromise.done { $0.open(urlString: openURLRaw) }
+            sceneManager.webViewWindowControllerPromise.done { $0.open(urlString: openURLRaw) }
         } else if let openURLDictionary = userInfo["url"] as? [String: String] {
             let url = openURLDictionary.compactMap { key, value -> String? in
                 if response.actionIdentifier == UNNotificationDefaultActionIdentifier,
@@ -1104,7 +870,7 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
             }.first
 
             if let url = url {
-                webViewWindowControllerPromise.done { $0.open(urlString: url) }
+                sceneManager.webViewWindowControllerPromise.done { $0.open(urlString: url) }
             } else {
                 Current.Log.error(
                     "couldn't make openable url out of \(openURLDictionary) for \(response.actionIdentifier)"
@@ -1174,39 +940,6 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
             let navController = UINavigationController(rootViewController: view)
             rootViewController?.present(navController, animated: true, completion: nil)
         })
-    }
-}
-
-enum XCallbackError: FailureCallbackError {
-    case generalError
-    case eventNameMissing
-    case serviceMissing
-    case templateMissing
-
-    var code: Int {
-        switch self {
-        case .generalError:
-            return 0
-        case .eventNameMissing:
-            return 1
-        case .serviceMissing:
-            return 2
-        case .templateMissing:
-            return 2
-        }
-    }
-
-    var message: String {
-        switch self {
-        case .generalError:
-            return L10n.UrlHandler.XCallbackUrl.Error.general
-        case .eventNameMissing:
-            return L10n.UrlHandler.XCallbackUrl.Error.eventNameMissing
-        case .serviceMissing:
-            return L10n.UrlHandler.XCallbackUrl.Error.serviceMissing
-        case .templateMissing:
-            return L10n.UrlHandler.XCallbackUrl.Error.templateMissing
-        }
     }
 }
